# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Pawnshop Management System API",
    "description": "A comprehensive API for managing pawnshop operations including loans, customers, items, payments, and sales.",
    "version": "1.0",
    "contact": {
      "name": "API Support",
      "email": "support@pawnshop.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8090",
  "basePath": "/api/v1",
  "schemes": ["http"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Type 'Bearer' followed by a space and the JWT token."
    }
  },
...
  "paths": {
...
    "/loans": {
      "get": {
        "tags": ["Loans"],
        "summary": "List Loans",
        "description": "Get a paginated list of loans",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "page", "in": "query", "type": "integer", "default": 1},
          {"name": "per_page", "in": "query", "type": "integer", "default": 20},
          {"name": "branch_id", "in": "query", "type": "integer"},
          {"name": "customer_id", "in": "query", "type": "integer"},
          {"name": "status", "in": "query", "type": "string"}
        ],
        "responses": {
          "200": {
            "description": "List of loans"
          }
        }
      },
      "post": {
        "tags": ["Loans"],
        "summary": "Create Loan",
        "description": "Create a new loan",
        "security": [{"BearerAuth": []}],
        "responses": {
          "201": {
            "description": "Loan created"
          }
        }
      }
    },
    "/payments": {
      "get": {
        "tags": ["Payments"],
        "summary": "List Payments",
        "description": "Get a paginated list of payments",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of payments"
          }
        }
      },
      "post": {
        "tags": ["Payments"],
        "summary": "Create Payment",
        "description": "Process a new payment",
        "security": [{"BearerAuth": []}],
        "responses": {
          "201": {
            "description": "Payment processed"
          }
        }
      }
    },
    "/sales": {
      "get": {
        "tags": ["Sales"],
        "summary": "List Sales",
        "description": "Get a paginated list of sales",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of sales"
          }
        }
      },
      "post": {
        "tags": ["Sales"],
        "summary": "Create Sale",
        "description": "Create a new sale",
        "security": [{"BearerAuth": []}],
        "responses": {
          "201": {
            "description": "Sale created"
          }
        }
      }
    },
...
    "/reports/dashboard": {
      "get": {
        "tags": ["Reports"],
        "summary": "Dashboard Stats",
        "description": "Get dashboard statistics",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "Dashboard statistics"
          }
        }
      }
    },
...
  },
...
}